+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.35.6
 
+* `Common`:
+    * `Version`:
+        * `Klock`: `2.3.1` -> `2.3.3`
+        * `MicroUtils`: `0.5.19` -> `0.5.21`
+
 ## 0.35.5
 
 **MIME TYPES FOR REQUESTS HAVE BEEN DEPRECATED DUE TO REDUNDANCY OF MIME TYPES IN FILES SENDING**